# internal/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the BengoBox treasury service.",
        "title": "Treasury Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Returns OK when the treasury API process is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Service liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.livenessResponse"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Exposes Prometheus metrics for scraping.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Prometheus metrics",
                "responses": {
                    "200": {
                        "description": "Prometheus metrics payload",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Validates connectivity to Postgres, Redis, NATS, and storage backends.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.readinessResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.readinessResponse"
                        }
                    }
                }
            }
        },
        "/v1/{tenantID}/ledger/chart-of-accounts": {
            "get": {
                "description": "Returns the ledger chart of accounts for the requesting tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger"
                ],
                "summary": "List chart of accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.chartOfAccountsResponse"
                        }
                    }
                }
            }
        },
        "/v1/{tenantID}/payments/intents": {
            "get": {
                "description": "Returns the payment intents that have been created for the tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "List payment intents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant identifier",
                        "name": "tenantID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.paymentIntentsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.chartOfAccountsResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ledgerAccount"
                    }
                }
            }
        },
        "handlers.ledgerAccount": {
            "type": "object",
            "properties": {
                "allowPostings": {
                    "type": "boolean",
                    "example": true
                },
                "balance": {
                    "type": "string",
                    "example": "0"
                },
                "code": {
                    "type": "string",
                    "example": "1000"
                },
                "currency": {
                    "type": "string",
                    "example": "KES"
                },
                "name": {
                    "type": "string",
                    "example": "Platform Cash"
                },
                "tenant": {
                    "type": "string",
                    "example": "tenant-123"
                },
                "type": {
                    "type": "string",
                    "example": "asset"
                }
            }
        },
        "handlers.livenessResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string",
                    "example": "treasury-app"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.paymentIntent": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "0.00"
                },
                "currency": {
                    "type": "string",
                    "example": "KES"
                },
                "id": {
                    "type": "string",
                    "example": "pi_123"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                }
            }
        },
        "handlers.paymentIntentsResponse": {
            "type": "object",
            "properties": {
                "intents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.paymentIntent"
                    }
                }
            }
        },
        "handlers.readinessResponse": {
            "type": "object",
            "properties": {
                "dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}